Add an event subscriber to set Turbolinks-Location headers on responses that follow a redirect
Because Turbolinks uses XHR requests, it only sees the final response in
the chain, with no way of knowing that the request has followed a
redirect. This means that the browser's address bar shows the URL that
was requested, not the correct URL for the response that was returned.

This MR adds an Event Subscriber which listens for Kernel
responses, sets the URL in the session on redirecting, and includes this
in the `Turbolinks-Location` header for the next response. On
seeing this, Turbolinks will silently update the URL so that the address
bar shows the correct URL.